+++ lib/utils.ts
@@ -111,7 +111,6 @@ export function removeNestedUndefinedValues(object: { [key: string]: any }): { [
  */
 export function optionSetup(ogsOptions: OpenGraphScraperOptions): { options: OpenGraphScraperOptions } {
   const options: OpenGraphScraperOptions = {
-    ogImageFallback: true,
     onlyGetOpenGraphInfo: false,
     ...ogsOptions,
   };