+++ tests/unit/utils.spec.js
@@ -222,15 +222,11 @@ describe('utils', function () {
   describe('optionSetup', function () {
     it('when passing nothing into optionSetup', function () {
       const { options } = utils.optionSetup();
-      expect(options).to.eql({ ogImageFallback: true, onlyGetOpenGraphInfo: false });
-    });
-    it('when passing ogImageFallback into optionSetup', function () {
-      const { options } = utils.optionSetup({ ogImageFallback: false });
-      expect(options).to.eql({ ogImageFallback: false, onlyGetOpenGraphInfo: false });
+      expect(options).to.eql({ onlyGetOpenGraphInfo: false });
     });
     it('when passing onlyGetOpenGraphInfo into optionSetup', function () {
       const { options } = utils.optionSetup({ onlyGetOpenGraphInfo: true });
-      expect(options).to.eql({ ogImageFallback: true, onlyGetOpenGraphInfo: true });
+      expect(options).to.eql({ onlyGetOpenGraphInfo: true });
     });
   });
 });