+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Adding `options.timeout` to set the fetch request timeout. (default is 10 seconds)
 - Remove `null` values from ogImages/ogVideos/twitterImages/twitterPlayers/musicSongs
 - Removing `options.allMedia`, you can just grab the first value of the array for the pervious behavior
+- Removing `options.ogImageFallback`, you can set `options.onlyGetOpenGraphInfo` to `true` for the pervious behavior
 - ogImages/ogVideos/twitterImages/twitterPlayers/musicSongs will always be an array now, you can just grab the first value of the array for the pervious behavior
 
 ## 5.2.3